refactor: move platform implementations to their folders
